First version of cursor routing.
Routing information is stored in the rt member of the translator.

Whenever the translate method is called with the optional arguments set to the index of the input in the original source and the index of the first character that will be written to the output the rt will be updated.

Next stage is to make the optional args propogate downwards to recursive calls to translate.