Rewrote a lot of the ueb translator, make it a whole lot simpler. I won't explain all the algorithums in detail -- it's much clearer than before and I've commented in the source. Most things should work as expected now. Things still to do: * handle lettersigns properly if a letter is next to a $ sign, this won't be very hard to do * Handle \{ and \} commands * Fix up a couple of the functions still dependent on the hold markup eg. simple fractions \frac{1}{2}
I Read through the patch in it's entirety and looks good to me. 

I've got no idea why I didn't look at this implementation to begin with
it makes the whole thing so much easier to deal with. Admittedly, it
wasn't that apparent and did require a bit of creativity :) 
Saved 187 lines of code :)