Add the ability to use custom .table files. i.e. You can overwrite our default translation strings for simple commands such as \pi, or add your own custom commands such as \implies.
This gives the advantage over the preprocessor of working better with
brltty and routing, as well as allowing us to change existing commands
and add new commands which may not be equivilant to any existing
commands. 
The preprocessor still serves an important purpose though, as I think it
works with more complex commands with arguments, where this will not.